requirements-dev.txt
pre-commit==3.5.0
isort==5.12.0
black==23.9.1
flake8==6.1.0
mypy==1.11.2
chromadb==0.5.11
transformers==4.45.1
torch==2.2.2
# torchvision==0.19.1
psycopg2-binary==2.9.9
